io_uring: fall back to a long user_data when the write-operation ids run out
Motivation:

Write operations are keyed by the 16 bits io_uring's user_data leaves once the registration id and the
opcode are packed into it, so there are only Short.MAX_VALUE of them. A zero-copy send holds its id until
the notification arrives rather than until the primary CQE, and notifications follow TCP ack processing,
so a peer that stops acking can keep every id outstanding. The submission then returned 0 and the write
waited for an id to come back instead of going out.

Modification:

nextZeroCopyWriteOperationId() takes from the short pool while it has room and otherwise hands out a
monotonic id above Short.MAX_VALUE. Such a value does not survive a round trip through short, so
IoUringIoHandler.canUseFastPath(...) rejects it and the submission goes down the slow path that already
carries the full user_data in PendingOpMap. Nothing new is needed on the handler side.

Exhaustion needs no probing: the free list being empty and the sequence being at its bound is exactly the
condition, and both are already tracked.

Ids above the bound cannot index the slot array, so they live in a map that is only allocated once the
pool has run out. The shutdown walk and the deregistration sweep cover it, and an entry is dropped as
soon as its operation completes or rolls back, since these ids are never recycled.

Write completions carry the user_data as a long from IoUringIoHandler down to writeComplete0. Reads,
polls, cancels and connects still narrow to short, as none of them can be given an overflow id.

Result:

A zero-copy write is submitted even when every short id is held by a send waiting for its notification.

Note: the fallback is deliberately not a probe. Scanning the id space to discover it is full would put a
65535 iteration loop on the event loop for the one case the escape hatch exists to handle.

+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/AbstractIoUringStreamChannel.java
@@ -699,7 +699,7 @@ abstract class AbstractIoUringStreamChannel extends AbstractIoUringChannel imple
         }
 
         @Override
-        boolean writeComplete0(byte op, int res, int flags, short data, int outstanding) {
+        boolean writeComplete0(byte op, int res, int flags, long data, int outstanding) {
             if ((flags & Native.IORING_CQE_F_NOTIF) == 0) {
                 // We only want to reset these if IORING_CQE_F_NOTIF is not set.
                 // If it's set we know this is only an extra notification for a write but we already handled
@@ -722,7 +722,10 @@ abstract class AbstractIoUringStreamChannel extends AbstractIoUringChannel imple
             Object current = channelOutboundBuffer.current();
             if (current instanceof IoUringFileRegion) {
                 IoUringFileRegion fileRegion = (IoUringFileRegion) current;
-                return handleWriteCompleteFileRegion(channelOutboundBuffer, fileRegion, res, data);
+                // Only zero-copy writes are ever handed an id outside the short range, and a splice is not one
+                // of them: IoUringFileRegion picks SPLICE_TO_PIPE / SPLICE_TO_SOCKET as its own data to tell
+                // its two stages apart, so narrowing here can not drop any bits.
+                return handleWriteCompleteFileRegion(channelOutboundBuffer, fileRegion, res, (short) data);
             }
 
             if (current instanceof FileRegion) {
